Moved a ton of public methods for ptypes to internal names to free up custom-methods. Deprecated support for long and cleaned up the code a little bit.
lib/ptypes:
    Fixed a bug in bitmap.{ror,rol} due to forgetting to clamp the value to the bit size.
    Added full support for array slicing to pbinary.array
        Implemented a few testcases to be sure.
    Renamed .get and .set to use internal names .__get__ and .__set__
        Implemented generic.{set,get} so that they point to the internal names in order to enable overloading.
    Modified all usages of .next() for iterators/generators to use python3's next keyword.
    Renamed .getindex to an internal name .__getindex__
    Removed the .walkto method due to it having never being used.
    Renamed .deserialize_block to an internal name .__deserialize_block__.
        Also renamed .deserialize_stream to .__deserialize_stream__.
    Removed a duplicate definition of pbinary.array.blockbits
    Renamed pbinary.partial.binaryobject to an internal name .__pb_object
    Added an alias for a pbinary.partial.object to .o
    Renamed pbinary.*.boffset back to pbinary.*.suboffset as in a previous commit.
        Added aliases for .suboffset in .bofs and .boffset just in case.
    Replaced all explicit testing of '.__class__' with proper usage of isinstance.
    Replaced all single-instances of 'str' with 'basestring'.
    Replaced all instances of explicit tests for '(int, long)' with six.integer_types.
    Removed any instance of .long or usage of the long() keyword and replaced it with int
    Renamed bitmap.number to bitmap.value, and added the aliases .number and .num.
    Replaced all usage of .num() with .int() and left an alias for it in .num() and .number()
    Fixed an uncaught exception in provider.lldb when trying to read 0 bytes.
    Implemented a function, bitmap.weight, to calculate the hamming-weight of a bitmap.
    Fixed a bug when trying to assign attributes to a pbinary which instead got
        assigned to the pbinary.partial instead of the intended element.
    Renamed .update_attributes to an internal name .__update___, and updated all relevant code.
    Modified the configuration by moving Config.display.partial out of hiding an into Config.pbinary.
        Replaced the booleans for hex and fractional with option fields.
    Fixed an mistake in bitmap.splitter that resulted in binaries being flipped.
        Changed all %-based format string to .format. Replaced all assertions with explicit tests.

lib/{pecoff,ndk}:
    Removed instances of .long() due to a change within this commit.

template:
    Fixed a few templates to remove the usage of the some internal naming in this commit.
    Updated archive.tar and iso9660 to remove an explicit usage of .long() which is removed in this commit.

+++ lib/ndk/ldrtypes.py
@@ -64,8 +64,8 @@ class LDR_DATA_TABLE_ENTRY(pstruct.type):
     ]
 
     def contains(self, address):
-        left = self['DllBase'].long()
-        right = left + self['SizeOfImage'].long()
+        left = self['DllBase'].int()
+        right = left + self['SizeOfImage'].int()
         return (address >= left) and (address < right)
 
 ## declarations, heh.
